+++ Hammer/Commands/AltCommand.cs
@@ -34,7 +34,7 @@ internal sealed class AltCommand : ApplicationCommandModule
         DiscordUser olderAccount = user.CreationTimestamp > alt.CreationTimestamp ? alt : user;
 
         var embed = new DiscordEmbedBuilder();
-        embed.WithAuthor(user.GetUsernameWithDiscriminator(), user.GetAvatarUrl(ImageFormat.Png));
+        embed.WithAuthor(user.GetUsernameWithDiscriminator(), iconUrl: user.GetAvatarUrl(ImageFormat.Png));
         embed.WithColor(DiscordColor.Green);
         embed.WithTitle("Alt account registered");
         embed.WithDescription("The following users have been registered as alts of each other.");
@@ -54,7 +54,7 @@ internal sealed class AltCommand : ApplicationCommandModule
         await _altAccountService.RemoveAltAsync(user, alt, context.Member).ConfigureAwait(false);
 
         var embed = new DiscordEmbedBuilder();
-        embed.WithAuthor(user.GetUsernameWithDiscriminator(), user.GetAvatarUrl(ImageFormat.Png));
+        embed.WithAuthor(user.GetUsernameWithDiscriminator(), iconUrl: user.GetAvatarUrl(ImageFormat.Png));
         embed.WithColor(DiscordColor.Orange);
         embed.WithTitle("Alt account unregistered");
         embed.WithDescription("The following users have been registered as alts of each other.");
@@ -73,7 +73,7 @@ internal sealed class AltCommand : ApplicationCommandModule
         IReadOnlyCollection<ulong> altAccounts = _altAccountService.GetAltsFor(user.Id);
 
         var embed = new DiscordEmbedBuilder();
-        embed.WithAuthor(user.GetUsernameWithDiscriminator(), user.GetAvatarUrl(ImageFormat.Png));
+        embed.WithAuthor(user.GetUsernameWithDiscriminator(), iconUrl: user.GetAvatarUrl(ImageFormat.Png));
         embed.WithColor(DiscordColor.Blurple);
         embed.WithTitle("Known alt accounts");