+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed an issue with the alt log embed mentioning the wrong user
+- Fixed an issue where the author of the alt embed was linked, rather than icon url'd
 
 ## [5.5.0] - 2023-06-09